requirements.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
pkg-resources==0.0.0
pytz==2018.4
